Remove sequential_reindexer Distro A, OPSEC #4584
