+++ rosbag2_cpp/CMakeLists.txt
@@ -61,8 +61,7 @@ add_library(${PROJECT_NAME} SHARED
   src/rosbag2_cpp/types/introspection_message.cpp
   src/rosbag2_cpp/writer.cpp
   src/rosbag2_cpp/writers/sequential_writer.cpp
-  src/rosbag2_cpp/reindexer.cpp
-  src/rosbag2_cpp/reindexers/sequential_reindexer.cpp)
+  src/rosbag2_cpp/reindexer.cpp)
 
 ament_target_dependencies(${PROJECT_NAME}
   PUBLIC
